Extracting Deparse::Interval code to own file.
